Code-audit Tier 3: module SW split, wire canonical, popover factory, publish prep
F1 — sw.js is now a module service worker (manifest "type": "module") and is
split into a DAG of small modules: state.mjs (-> wire.mjs) <- worktab.mjs <-
overlay.mjs (-> theme-tokens.js) <- actions.mjs; port.mjs -> {state, wire,
overlay, actions}; panel-api.mjs -> {state, port, overlay}; sw.js ->
{port, panel-api}. Behavior-identical; verified by the full battery + the
sw-e2e module-realm boot (vm.SourceTextModule) and real-Chrome suites.

F2 — browser tool DOM side lands in dom-actions.js behind the
__dshAugDom global (overlay injects it); the 'html' wire action is dropped
(the five browser_* tools cover it).

F5 — repo-root wire.mjs is the canonical protocol module (encode/decode,
genId, Pending); extension/wire.mjs is a byte-identical copy, asserted by
sw-e2e at boot (Buffer.equals). pipe.mjs now consumes Pending properly —
the battery caught a residual Map-style plugin-waiter that threw on every
augmentor/* request (save) and left a stale timer; fixed via
Pending.add/settle.

F6 — theme-tokens.js is the single source of palette/tokens; veil.js,
lab/veil-preview.html and the overlay read from it.

F10 — the header's three popovers (model picker, sessions, hue) are built
on one popover() factory (sidepanel.js): open/close/position/toggle,
outside-mousedown + Escape close, optional .open class.

D3 — publish metadata: root LICENSE (MIT), plugin/LICENSE, plugin/README.md,
and the built plugin/dist/ is committed so the live app can hot-reload
without a build step (dist is not git-ignored).

D4 — plugin pnpm >=10 allowBuilds: pnpm-workspace.yaml + lockfile refresh.

D5 — install-native-host.sh now guards against double mounts: if the live
patch file holds exactly one ?src= extension mount and >=1 bare-name
mount, the M4 bare-name block is dropped (with .bak + stderr warning) so a
published install never shadows the live one. README gains a Security
posture section (fence rationale, token channel, 0600 file, config wins).

e2e (fix-not-delete): the marker-session row oracles in panel/chrome/m2
e2e matched rows by rendered title, which races the DSH app's async
re-titling (marker title is replaced by the assistant's first line after
the turn). Rows now carry data-session-id and the suites pin the new
session via a session.list id diff. panel-e2e's final assert now checks
rendered user + assistant prose instead of a 200-char log floor that
depended on model verbosity.

Battery (all green): sw, panel, m3, chrome, m2, tools e2e + plugin boot
(fresh-copy, token-gated).
