+++ extension/manifest.json
@@ -1,12 +1,12 @@
 {
   "manifest_version": 3,
   "name": "Augmentor Agent powered by DSH",
-  "version": "0.1.25",
+  "version": "0.1.26",
   "description": "Augmentor: a DSH-powered sidecar agent that controls this browser via the native-messaging bridge.",
   "key": "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAyVanGT7jyrPt4GkeHVSdwviET907D/MOg2WF5dtkGSNYs0mN7F10owyQAf0P1X0v7EsgC9zz7YVyUAFLxrWjH9iDNCslPUPThWjRz0G6ZgR/oFNu2gEv+o8XVgl+y/NZhsD6Y9yJV3uJDSeCHVHRa+KCZ5L3bZx6OKCLVuj2e5pjqE3V/1su8LgtyUZern4il7haxA4Tp4wW6BtSEKXu9raGFlw2JRczagbMxqA7kSaehzG/dq/avUKRWLy+iD7e1eNyYyU4uWs1v9rlkzJbhveObJkINVC8rW0oBGJv4so2eAk3MklmW9zbB71ZHK3Fl94nBB7n2DqRh9qO6PCg5QIDAQAB",
   "permissions": ["nativeMessaging", "tabs", "scripting", "activeTab", "sidePanel", "windows", "storage"],
   "host_permissions": ["http://*/*", "https://*/*"],
-  "background": { "service_worker": "sw.js" },
+  "background": { "service_worker": "sw.js", "type": "module" },
   "side_panel": { "default_path": "sidepanel.html" },
   "action": {
     "default_title": "Augmentor — open side panel",
